Move optimization code to another file
